requirements.txt

docopt==0.6.2
ipaddress==1.0.22
pykwalify==1.6.1
python-dateutil==2.7.3
PyYAML==3.12
six==1.11.0

